+++ src/KilnUniV3.t.sol
@@ -84,7 +84,7 @@ contract KilnTest is Test {
     function mintDai(address usr, uint256 amt) internal {
         hevm.store(
             address(DAI),
-            keccak256(abi.encode(address(usr), uint(2))),
+            keccak256(abi.encode(address(usr), uint256(2))),
             bytes32(uint256(amt))
         );
         assertEq(GemLike(DAI).balanceOf(address(usr)), amt);
